Tests pass on all platforms with and without phone
